fix(chatbot): decouple response-example contracts
